split JsonConfigLoader into a loader and parser class. Dto classes now implement JsonSerializable interface
